A very very rough draft
This in development piece of code should help with analyzing physiological triggers (identification, segmentation and naming).
The naming section is based on the setup in our lab, where we use an online google sheet to write the order of the tasks and the number of TR given during the experiment.
That said, you probably have a different setup in regards to naming tasks, so that part won't be as beneficial to you now. We are working on making that part automatic using a biopac TTL adapter. Good developments coming soon.